requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
anthropic==0.40.0
openai==1.56.2
httpx==0.27.2
aiofiles==24.1.0
python-multipart==0.0.12
jinja2==3.1.4
sse-starlette==2.1.3
pydantic-settings==2.6.1
requests==2.32.3
youtube-transcript-api==1.2.4
